+++ Tarea5_RMI/Servidor/InterfaceRMI.java
@@ -2,6 +2,6 @@
 import java.rmi.*;
 
 public interface InterfaceRMI extends Remote {
-    double[][] multiplica_matrices(double[][] A, double[][] B,int N) throws RemoteException;
+    float[][] multiplica_matrices(float[][] A, float[][] B,int N) throws RemoteException;
     
 }
